requirements.txt
Django==5.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
SpeechRecognition
beautifulsoup4
dash==2.14.2
numpy==1.26.2
pandas==2.1.4
pillow==10.2.0
plotly==5.18.0
validate-email==1.3
xhtml2pdf==0.2.13
reportlab==4.0.9
requests==2.31.0